trying to implement current nodes
