+++ main.py
@@ -3,7 +3,21 @@ stations_data = loadStationsData()
 
 from playerLogic import *
 
-from gameLogic import *
+def updateGameDetectives(detectives):
+    for detective in detectives:
+        detective.setNewNode()
+        printSeparator()
+
+def updateGameMrX(mr_x):
+    print("MR X'S TURN!!!")
+    if mr_x.double_move_tickets>1:
+        mr_x.askForDoubleMove()
+    mr_x.setNewNode()
+
+def printSeparator():
+    print()
+    print("*"*50)
+    print()
 
 
 def main():
@@ -54,15 +68,16 @@ def main():
                 else:
                     continue
             printSeparator()
+
+    #print(detective_nodes)
     
     # print(active_nodes)
-    
 
-
-    # printSeparator()
-    # updateGameMrX(mr_x)
-    # printSeparator()
-    # updateGameDetectives(detectives)
+    while active_nodes[0] not in active_nodes[1:len(active_nodes)]:
+        printSeparator()
+        updateGameMrX(mr_x)
+        printSeparator()
+        updateGameDetectives(detectives)
 
 if __name__ == "__main__":
     main()
